pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[project]
name = "pyrotation"
version = "0.0.2"
authors = [
    { name = "fzhiheng", email = "fzhazr@sjtu.edu.cn" },
]
maintainers = [
    { name = "fzhiheng", email = "fzhazr@sjtu.edu.cn" },
]
description = "A package to handle rotation in 3D space with pytorch and numpy"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "torch",
    "plum-dispatch",
]
keywords = ["rotation", "pytorch", "numpy", "quaternion", "euler", "axis-angle", "so3", "se3", "SO3", "SE3", "lie algebra"]

[project.urls]
Homepage = "https://github.com/fzhiheng/pyrotation.git"
